Moved errors functions into different files
